feat: Fix typo in gaming.nix
This commit fixes a typo in the gaming.nix file. The word "gamimg" has been corrected to "gaming" in the config.programs.steam section. This ensures that the correct profile is enabled for gaming and opens the necessary ports in the firewall for Steam Remote Play and Source Dedicated Server.