feat: Add support for macOS profiles
This commit adds support for macOS profiles by introducing the `darwin.nix` file and updating the necessary configurations in `default.nix` and `configuration.nix`. Now, users can define their macOS-specific settings and packages in the `darwin.nix` file, and the system will include them when the `darwin` profile is enabled.